refactor: update package name at document mapping
