+++ jnosql-mapping/jnosql-mapping-column/src/main/java/org/eclipse/jnosql/mapping/column/query/DynamicQuery.java
@@ -17,10 +17,10 @@ package org.eclipse.jnosql.mapping.column.query;
 import jakarta.data.repository.Limit;
 import jakarta.data.repository.Sort;
 import org.eclipse.jnosql.communication.column.ColumnQuery;
-import org.eclipse.jnosql.mapping.NoSQLPage;
+import org.eclipse.jnosql.mapping.core.NoSQLPage;
 import org.eclipse.jnosql.mapping.column.MappingColumnQuery;
 import org.eclipse.jnosql.mapping.core.repository.DynamicReturn;
-import org.eclipse.jnosql.mapping.repository.SpecialParameters;
+import org.eclipse.jnosql.mapping.core.repository.SpecialParameters;
 
 import java.util.ArrayList;
 import java.util.List;
